[refact] modal, flatbutton component 수정
