+++ src/page/HomeComponents/Calendar.js
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import { useDate } from "../../contexts/dateProvider";
+import Modal from "../../components/Modal";
 import TextButton from "../../components/TextButton";
 import CalendarContents from "./CalendarContents";
 
@@ -12,17 +13,19 @@ export default function Calendar({ setOpenCalendar }) {
   };
 
   return (
-    <div className="modalBackground">
-      <div className="modal" style={{ height: "45vh", padding: "0 10px" }}>
-        <CalendarContents
-          selectedDate={selectedDate}
-          setSelectedDate={setSelectedDate}
-        />
-        <div className="textBtnDiv" style={{ marginTop: "15px" }}>
-          <TextButton onClick={handleChangeDate} text="확인" />
-          <TextButton onClick={() => setOpenCalendar(false)} text="취소" />
-        </div>
+    <Modal
+      buttonType="text"
+      style={{ height: "45vh", padding: "0 10px" }}
+      handleModalClick={() => setOpenCalendar(false)}
+    >
+      <CalendarContents
+        selectedDate={selectedDate}
+        setSelectedDate={setSelectedDate}
+      />
+      <div className="textBtnDiv" style={{ marginTop: "15px" }}>
+        <TextButton onClick={handleChangeDate} text="확인" />
+        <TextButton onClick={() => setOpenCalendar(false)} text="취소" />
       </div>
-    </div>
+    </Modal>
   );
 }
