+++ src/style/index.scss
@@ -413,7 +413,7 @@ $background-color: #fbf8f3;
   border-radius: 10px;
   background-color: white;
   width: 80vw;
-  height: 40vh;
+  height: 25vh;
 
   .modalMessage {
     position: fixed;
